pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "finansy-url-shortener"
version = "0.1.0"
dependencies = [
  "aiosqlite",
  "envyaml",
  "fastapi",
  "pydantic",
  "PyYAML",
  "SQLAlchemy"
]
requires-python = ">= 3.11"
readme = "README.md"